requirements.txt
pandas
openpyxl
requests
selenium
beautifulsoup4
lxml
python-dotenv
xlsxwriter
schedule
pytest
pytest-cov
streamlit
matplotlib
aiohttp>=3.8.0
asyncio
